Make compute lock be more fine grained for start
On container activate we may need to pull an image or wait for container
IP to be assigned.  During this time we do not want to block other
container operations.
